requirements_test_all.txt
# Open Peer Power tests, full dependency set
# Automatically generated by gen_requirements_all.py, do not edit

-r requirements_test.txt

# openpeerpower.components.mobile_app
PyNaCl==1.3.0

# openpeerpower.auth.mfa_modules.totp
PyQRCode==1.2.1

# openpeerpower.components.automatic
aioautomatic==0.6.5

# openpeerpower.components.http
aiohttp_cors==0.7.0

# openpeerpower.components.zha
bellows-openpeerpower==0.13.2

# openpeerpower.components.bom
bomradarloop==0.1.3

# openpeerpower.scripts.check_config
colorlog==4.1.0

# openpeerpower.scripts.credstash
# credstash==1.15.0

# openpeerpower.components.updater
distro==1.4.0

# openpeerpower.components.season
ephem==3.7.7.0

# openpeerpower.components.mqtt
hbmqtt==0.9.5

# openpeerpower.scripts.keyring
keyring==20.0.0

# openpeerpower.scripts.keyring
keyrings.alt==3.4.0

# openpeerpower.components.discovery
# openpeerpower.components.ssdp
netdisco==2.6.0

# openpeerpower.components.zwave
openpeerpower-pyozw==0.1.8

# openpeerpower.components.cloud
opp-nabucasa==0.31

# openpeerpower.components.mqtt
paho-mqtt==1.5.0

# openpeerpower.components.met
pyMetno==0.4.6

# openpeerpower.components.almond
pyalmond==0.0.2

# openpeerpower.components.daikin
pydaikin==1.6.2

# openpeerpower.components.zwave
pydispatcher==2.0.5

# openpeerpower.components.modbus
pymodbus==1.5.2

# openpeerpower.auth.mfa_modules.notify
# openpeerpower.auth.mfa_modules.totp
# openpeerpower.components.otp
pyotp==2.3.0

# openpeerpower.components.recorder
sqlalchemy==1.3.11

# openpeerpower.components.yr
xmltodict==0.12.0

# openpeerpower.components.zeroconf
zeroconf==0.24.4

# openpeerpower.components.zha
zha-quirks==0.0.33

# openpeerpower.components.zha
zigpy-cc==0.1.0

# openpeerpower.components.zha
zigpy-deconz==0.7.0

# openpeerpower.components.zha
zigpy-openpeerpower==0.13.2

# openpeerpower.components.zha
zigpy-xbee-openpeerpower==0.9.0

# openpeerpower.components.zha
zigpy-zigate==0.5.1
